Asercom protocol support cont'd
* asercom protocol almost complete (missing commands "-U" and "R")
* extend asercom to include additional commands for the new sensors
available on e-puck2:
- this protocol is called asercom2, available on selector position 8
- support for 4 microphones and distance sensor
* e-puck1.x library:
- maintain same structure as with e-puck1.x library, instead of one
single file
- extended support in order to run the asercom protocol
* moved selector handling in its own module file instead of "utility"
* extended imu module:
- gyro calibration
- acc and gyro get functions
*  distance sensor module:
- added "get distance in mm" function

+++ src/audio/microphone.c
@@ -18,7 +18,7 @@
  *       ###
  *
  */
-int32_t mic_volume[4];
+uint16_t mic_volume[4];
 int16_t mic_last[4];
 
 int16_t mic_get_last(uint8_t mic) {
@@ -29,7 +29,7 @@ int16_t mic_get_last(uint8_t mic) {
 	}
 }
 
-int32_t mic_get_volume(uint8_t mic) {
+uint16_t mic_get_volume(uint8_t mic) {
 	if(mic < 4) {
 		return mic_volume[mic];
 	} else {
