+++ src/audio/microphone.h
@@ -6,8 +6,10 @@
 extern "C" {
 #endif
 
+#include <hal.h>
+
 void mic_start(void);
-int32_t mic_get_volume(uint8_t mic);
+uint16_t mic_get_volume(uint8_t mic);
 int16_t mic_get_last(uint8_t mic);
 
 #ifdef __cplusplus
